backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.23
numpy>=1.24.3
sentence-transformers>=2.2.2
requests
email-validator
psycopg2-binary>=2.9.9